chore: code style change, lint with algolia-standard
Following #128 and various discussions we already had on code style, I
implemented `algolia-standard` in the project.

The goal of this PR is to provide a solid ground for an easy linter
that we can use on all our projects.

The currently picked coding style is the one from Airbnb, with
modifications.

You can find the rules here: https://github.com/algolia/standard#rules

We can discuss the rules of course as long as we do not discuss them
every week!

Other projects:

- Original feross/standard:
After trying it I discovered some rules were not strict enough and
would lead to again style discussions. Has some sort of formatter but
not very good.

- JSCS:
I tried JSCS: good but only oriented on style, no rules on unused
variables or ways to prevent bugs like eslint does. Has some sort of
formatter, but not very good.

Ultimately eslint will have a good formatter, in the meantime we will
all switch to some new coding style, for good :-)?

fixes #128