docs(jsdoc): fix spelling
I have spelling detection enabled in my editor, and it complained that there were a few typos, so I fixed them.

There's one typo I can't really fix or I'm not sure if it's worth it since it changes a parameter.

There's `...Sizefor...` in a few attributes that should be `...SizeFor...`.