feat(facets): add expandWildcardFacets parameter to process * in facets
When * is present in facets, we don't need to request the individual facet values of each facet that's declared, as * will already request it.

With a new parameter "expandWildcardFacets", we filter out all facets that aren't * itself or -something. The new parameter doesn't add "*" automatically, it still

To decide:
- is "-something" worth caring about?
- when * isn't present, should no facets be requested or the default ones?
- is this the right location, maybe the client cache should have similar logic instead?

see: https://github.com/algolia/algoliasearch-helper-js/issues/391