fix(requests): send a sorted object of parameters
Having a sorted set of parameters helps with making sure the cache is hit in more cases.

For example, when you remove and immediately add the same widget again, and they no longer are in the same order, but have the same parameters conceptually.

This PR does not sort the items inside the parameters, eg. facets, which would be possible, but isn't done yet.

FX-1473