fix(getFacetValues): reflect the value of _state in hierarchicalFacetValues (#925)
Before this commit, the behaviour of conjunctive/disjunctive facets and hierarchical facets is different when `results._state` has changed since the instantiation of the SearchResults class.

Conjunctive/disjunctive take the current value of `_state`, while `hierarchical` took `isRefined` from the constructor.

We want to make this change to enable optimistic UI, where we render the existing results with the next state. This causes the `isRefined` state to immediately be useful for display.