docs(html): no indentation after markdown-in-html
Having the tag after markdown be directly a dif, but indented causes the tags not to properly be closed somehow. Not sure if this is expected, or an engine bug, but now at least all tags are closed.

I'm not sure if long-term we're spending much effort here, but at least it renders correctly now.

test path: check reference.html in a html validator